+++ R/rows.R
@@ -145,11 +145,14 @@ rows_insert.data.frame <- function(x,
 
   by <- rows_check_by(by, y)
 
-  rows_check_containment(x, y)
+  rows_check_x_contains_y(x, y)
+  rows_check_contains_by(x, by, "x")
+  rows_check_contains_by(y, by, "y")
+
   y <- rows_cast_y(y, x)
 
-  x_key <- rows_select_key(x, by, "x")
-  y_key <- rows_select_key(y, by, "y")
+  x_key <- dplyr_col_select(x, by)
+  y_key <- dplyr_col_select(y, by)
 
   keep <- rows_check_y_conflict(x_key, y_key, conflict)
 
@@ -181,7 +184,7 @@ rows_append.data.frame <- function(x,
 
   y <- auto_copy(x, y, copy = copy)
 
-  rows_check_containment(x, y)
+  rows_check_x_contains_y(x, y)
   y <- rows_cast_y(y, x)
 
   rows_bind(x, y)
@@ -214,18 +217,23 @@ rows_update.data.frame <- function(x,
 
   by <- rows_check_by(by, y)
 
-  rows_check_containment(x, y)
+  rows_check_x_contains_y(x, y)
+  rows_check_contains_by(x, by, "x")
+  rows_check_contains_by(y, by, "y")
 
-  x_key <- rows_select_key(x, by, "x")
-  y_key <- rows_select_key(y, by, "y", unique = TRUE)
+  x_key <- dplyr_col_select(x, by)
+  y_key <- dplyr_col_select(y, by)
+
+  rows_check_unique(y_key, "y")
+
   args <- vec_cast_common(x = x_key, y = y_key)
   x_key <- args$x
   y_key <- args$y
 
   values_names <- setdiff(names(y), names(y_key))
 
-  x_values <- x[values_names]
-  y_values <- y[values_names]
+  x_values <- dplyr_col_select(x, values_names)
+  y_values <- dplyr_col_select(y, values_names)
   y_values <- rows_cast_y(y_values, x_values)
 
   keep <- rows_check_y_unmatched(x_key, y_key, unmatched)
@@ -278,18 +286,23 @@ rows_patch.data.frame <- function(x,
 
   by <- rows_check_by(by, y)
 
-  rows_check_containment(x, y)
+  rows_check_x_contains_y(x, y)
+  rows_check_contains_by(x, by, "x")
+  rows_check_contains_by(y, by, "y")
 
-  x_key <- rows_select_key(x, by, "x")
-  y_key <- rows_select_key(y, by, "y", unique = TRUE)
+  x_key <- dplyr_col_select(x, by)
+  y_key <- dplyr_col_select(y, by)
+
+  rows_check_unique(y_key, "y")
+
   args <- vec_cast_common(x = x_key, y = y_key)
   x_key <- args$x
   y_key <- args$y
 
   values_names <- setdiff(names(y), names(y_key))
 
-  x_values <- x[values_names]
-  y_values <- y[values_names]
+  x_values <- dplyr_col_select(x, values_names)
+  y_values <- dplyr_col_select(y, values_names)
   y_values <- rows_cast_y(y_values, x_values)
 
   keep <- rows_check_y_unmatched(x_key, y_key, unmatched)
@@ -347,18 +360,23 @@ rows_upsert.data.frame <- function(x,
 
   by <- rows_check_by(by, y)
 
-  rows_check_containment(x, y)
+  rows_check_x_contains_y(x, y)
+  rows_check_contains_by(x, by, "x")
+  rows_check_contains_by(y, by, "y")
 
-  x_key <- rows_select_key(x, by, "x")
-  y_key <- rows_select_key(y, by, "y", unique = TRUE)
+  x_key <- dplyr_col_select(x, by)
+  y_key <- dplyr_col_select(y, by)
+
+  rows_check_unique(y_key, "y")
+
   args <- vec_cast_common(x = x_key, y = y_key)
   x_key <- args$x
   y_key <- args$y
 
   values_names <- setdiff(names(y), names(y_key))
 
-  x_values <- x[values_names]
-  y_values <- y[values_names]
+  x_values <- dplyr_col_select(x, values_names)
+  y_values <- dplyr_col_select(y, values_names)
   y_values <- rows_cast_y(y_values, x_values)
 
   loc <- vec_match(x_key, y_key)
@@ -413,8 +431,12 @@ rows_delete.data.frame <- function(x,
 
   by <- rows_check_by(by, y)
 
-  x_key <- rows_select_key(x, by, "x")
-  y_key <- rows_select_key(y, by, "y")
+  rows_check_contains_by(x, by, "x")
+  rows_check_contains_by(y, by, "y")
+
+  x_key <- dplyr_col_select(x, by)
+  y_key <- dplyr_col_select(y, by)
+
   args <- vec_cast_common(x = x_key, y = y_key)
   x_key <- args$x
   y_key <- args$y
@@ -470,7 +492,7 @@ rows_check_by <- function(by, y, ..., error_call = caller_env()) {
   by
 }
 
-rows_check_containment <- function(x, y, ..., error_call = caller_env()) {
+rows_check_x_contains_y <- function(x, y, ..., error_call = caller_env()) {
   check_dots_empty()
 
   bad <- setdiff(names(y), names(x))
@@ -493,43 +515,42 @@ rows_cast_y <- function(y, x, ..., call = caller_env()) {
   vec_cast(x = y, to = x, x_arg = "y", to_arg = "x", call = call)
 }
 
-rows_select_key <- function(x,
-                            by,
-                            arg,
-                            ...,
-                            unique = FALSE,
-                            error_call = caller_env()) {
+rows_check_contains_by <- function(x, by, arg, ..., error_call = caller_env()) {
   check_dots_empty()
 
   missing <- setdiff(by, names(x))
 
-  if (!is_empty(missing)) {
-    missing <- err_vars(missing)
-
-    message <- c(
-      "All columns specified through `by` must exist in `x` and `y`.",
-      i = glue("The following columns are missing from `{arg}`: {missing}.")
-    )
-
-    abort(message, call = error_call)
+  if (is_empty(missing)) {
+    return(invisible())
   }
 
-  out <- x[by]
+  missing <- err_vars(missing)
 
-  if (unique && vec_duplicate_any(out)) {
-    duplicated <- vec_duplicate_detect(out)
-    duplicated <- which(duplicated)
-    duplicated <- err_locs(duplicated)
+  message <- c(
+    "All columns specified through `by` must exist in `x` and `y`.",
+    i = glue("The following columns are missing from `{arg}`: {missing}.")
+  )
 
-    message <- c(
-      glue("`{arg}` key values must be unique."),
-      i = glue("The following rows contain duplicate key values: {duplicated}.")
-    )
+  abort(message, call = error_call)
+}
 
-    abort(message, call = error_call)
+rows_check_unique <- function(x, arg, ..., error_call = caller_env()) {
+  check_dots_empty()
+
+  if (!vec_duplicate_any(x)) {
+    return(invisible())
   }
 
-  out
+  duplicated <- vec_duplicate_detect(x)
+  duplicated <- which(duplicated)
+  duplicated <- err_locs(duplicated)
+
+  message <- c(
+    glue("`{arg}` key values must be unique."),
+    i = glue("The following rows contain duplicate key values: {duplicated}.")
+  )
+
+  abort(message, call = error_call)
 }
 
 rows_check_y_conflict <- function(x_key,